requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pymysql 
celery 
redis 
requests 
python-dotenv
cryptography